+++ lib/generators/cms/component/form_builder/templates/app/cells/form_builder_cell.rb
@@ -39,6 +39,8 @@ class FormBuilderCell < Cell::Rails
       label: @title,
       input_html: {
         maxlength: @attribute.attributes['max_length'],
+        rows: 10,
+        class: 'span5',
       },
       required: @mandatory,
     }