+++ lib/generators/cms/kickstart/templates/app/cells/footer_cell.rb
@@ -8,7 +8,16 @@ class FooterCell < Cell::Rails
     ]
   end
 
-  def show
+  def show(page)
+    @page = page
+    homepage = page.homepage
+
+    @footer_links = if homepage
+      homepage.footer_links
+    else
+      RailsConnector::LinkList.new(nil)
+    end
+
     render
   end