+++ lib/generators/cms/kickstart/templates/create_structure.rb
@@ -10,7 +10,7 @@ class CreateStructure < ::RailsConnector::Migration
 
     try_create_obj(_path: "<%= website_path %>", _obj_class: 'Website')
 
-    try_create_obj(_path: "<%= homepage_path %>", _obj_class: 'Homepage', headline: 'Homepage', locale: 'en')
+    try_create_obj(_path: "<%= homepage_path %>", _obj_class: 'Homepage', locale: 'en', headline: 'Company, Inc.')
 
     try_create_obj(_path: "<%= configuration_path %>", _obj_class: 'Container')
 
@@ -23,6 +23,13 @@ class CreateStructure < ::RailsConnector::Migration
       Obj.find_by_path("<%= homepage_path %>").id,
       error_not_found_page_link: [{ url: "<%= configuration_path %>/error-not-found" }],
       login_page_link: [{ url: "<%= configuration_path %>/login" }],
+      footer_links: [
+        { title: 'Getting Started', url: 'https://dev.infopark.net/getting-started' },
+        { title: 'Developer Guide', url: 'https://dev.infopark.net/developer-guide' },
+        { title: 'API Reference', url: 'https://dev.infopark.net/api-doc' },
+        { title: 'Blog', url: 'https://dev.infopark.net/blog'},
+        { title: 'Support', url: 'https://dev.infopark.net/support' },
+      ]
     )
 
     try_create_obj(_path: "<%= resources_path %>", _obj_class: 'Container', headline: 'Resources')